[refactor] move all python modules into src and test
