--- scraper/requirements.txt
@@ -1,9 +1,0 @@
-cycler==0.10.0
-decorator==4.4.1
-kiwisolver==1.1.0
-matplotlib>=2.0.0,<=3.1.1
-networkx>=2.2,<=2.4
-numpy>=1.15.0,<=1.18.1
-pyparsing==2.4.6
-python-dateutil==2.8.1
-six==1.14.0
