+++ src/iok/README.md
@@ -1,0 +1,8 @@
+# IoK Python Library
+
+Main IoK python utils library. 
+
+* `constants` contains some constants for generating Awesome Lists
+* `meta` contains meta-objects used by the IoK to represent Awesome Lists, Graphs, etc
+* `types` contains some reusable types
+*
